+++ src/main/java/com/example/gistcompetitioncnserver/config/LoginUserArgumentResolver.java
@@ -1,5 +1,8 @@
-package com.example.gistcompetitioncnserver.user;
+package com.example.gistcompetitioncnserver.config;
 
+import com.example.gistcompetitioncnserver.config.annotation.LoginUser;
+import com.example.gistcompetitioncnserver.user.SessionUser;
+import com.example.gistcompetitioncnserver.user.UserService;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
