+++ src/main/java/com/example/gistcompetitioncnserver/config/annotation/LoginRequired.java
@@ -1,12 +1,11 @@
-package com.example.gistcompetitioncnserver.user;
-
+package com.example.gistcompetitioncnserver.config.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.PARAMETER})
+@Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LoginUser {
+public @interface LoginRequired {
 }
