+++ src/main/java/com/example/gistcompetitioncnserver/config/annotation/LoginUser.java
@@ -1,11 +1,12 @@
-package com.example.gistcompetitioncnserver.user;
+package com.example.gistcompetitioncnserver.config.annotation;
+
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.METHOD})
+@Target({ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LoginRequired {
+public @interface LoginUser {
 }
